Reorganize MT Relay a bit
