+++ src/Domain/LeanCode.DomainModels.MassTransitRelay/Middleware/EventsPublisherElement.cs
@@ -6,7 +6,7 @@ using LeanCode.DomainModels.EventsExecution;
 using LeanCode.DomainModels.Model;
 using LeanCode.Pipelines;
 
-namespace LeanCode.DomainModels.MassTransitRelay
+namespace LeanCode.DomainModels.MassTransitRelay.Middleware
 {
     public class EventsPublisherElement<TContext, TInput, TOutput> : IPipelineElement<TContext, TInput, TOutput>
         where TContext : notnull, ICorrelationContext, IEventsInterceptorContext
