+++ src/Domain/LeanCode.DomainModels.MassTransitRelay/Middleware/EventsPublisherFilter.cs
@@ -6,7 +6,7 @@ using MassTransit;
 using MassTransit.ConsumeConfigurators;
 using MassTransit.PipeConfigurators;
 
-namespace LeanCode.DomainModels.MassTransitRelay
+namespace LeanCode.DomainModels.MassTransitRelay.Middleware
 {
     public class EventsPublisherFilter<TConsumer, TMessage> : IFilter<ConsumerConsumeContext<TConsumer, TMessage>>
         where TConsumer : class
