+++ src/Domain/LeanCode.DomainModels.MassTransitRelay/Middleware/StoreAndPublishEventsFilter.cs
@@ -6,7 +6,7 @@ using MassTransit;
 using MassTransit.ConsumeConfigurators;
 using MassTransit.PipeConfigurators;
 
-namespace LeanCode.DomainModels.MassTransitRelay
+namespace LeanCode.DomainModels.MassTransitRelay.Middleware
 {
     public class StoreAndPublishEventsFilter<TConsumer, TMessage> : IFilter<ConsumerConsumeContext<TConsumer, TMessage>>
         where TConsumer : class
