+++ test/Domain/LeanCode.DomainModels.MassTransitRelay.Tests/Integration/TestApp.cs
@@ -7,11 +7,11 @@ using LeanCode.Correlation;
 using LeanCode.CQRS.Default;
 using LeanCode.CQRS.Execution;
 using LeanCode.DomainModels.EventsExecution;
+using LeanCode.DomainModels.MassTransitRelay.Middleware;
 using LeanCode.DomainModels.Model;
 using LeanCode.IdentityProvider;
 using MassTransit;
 using Microsoft.Data.Sqlite;
-using Microsoft.EntityFrameworkCore;
 using Xunit;
 
 namespace LeanCode.DomainModels.MassTransitRelay.Tests.Integration
